+++ CHANGELOG.md
@@ -1,11 +1,23 @@
-# 2.0.1
+# 2.0.6, 2.0.7
 
-- Fixed Battleground-Check
+- Remove GreenWall from dependencies and check ingame if it is enabled
+
+# 2.0.5
+
+- Fixing false positive on "Alliance" NPCs
+
+# 2.0.4
+
+- Fixing DeathCOunt initialization and PvP Flag on neutral NPCs
+
+# 2.0.3
+
+- Remove Version hint for non for players without invite permissions
 
 # 2.0.2
 
 - Removed guild request message for players without invite permissions
 
-# 2.0.3
+# 2.0.1
 
-- Remove Version hint for non for players without invite permissions
+- Fixed Battleground-Check